Scheduler audit fixes 7494028250406511254

+++ headers/private/system/scheduler_defs.h
@@ -64,7 +64,7 @@ struct scheduling_analysis {
 
 struct loadavg {
 	uint32	ldavg[3];
-	long 	fscale;
+	int32 	fscale;
 };
